MPI: broadcast pt2.4
Al, in, ps, and de the allocatable arrays in 'MCvars'.
I did not perform this for a few arrays because they
are defined within the parallel part of the problem.
I did not do this for a few others because they tally
material-dependent flux, and some code regarding this
is missing, so that these variables are no longer
allocated.  If I plan to do things with binary media
I should figure out where this is and add it back.