LN Proc: SC, simpler input
I put the SC orders for an anisotropic problem in the same line
as for an isotropic problem.  This makes the input much simpler
to satisfy as a user.  You still have to keep a couple things
straight, but it is easier.  Namely, you must have declared
enough anisotropic orders as if you are going to use them
whether or not you do.

I also added a test for this input and fiddled with a couple
of other things.