Simplify-MPI: Debugged 'binary' geomtype
1) When removing WAMC I had removed a code piece that LPMC
also used.  Here I restored it.

2) My relatively new test for particle history convergence
is done when the number of particles reaches the max number
of particles.  This was being loaded from realization input
for LPMC and atmixMC, so that if the number of particles
chosen for one of these was greater than the ceiling
for the others in the input file, the algorithm would run
on an infinite loop.

3) Adamscases did not load a sourceType of 'leftiso'.
Since the input file did have 'leftbeam', the Adamscases
solves were all a little off.  Now 'leftiso' is loaded,
and we appear to be working correctly.

4) I made LPMC and atmixMC leakage value output formatted
in a better way.