LN Proc: 'correlated''anticorrelated''independent'
So it turns out that the three simple ways to handle two
cross sections as as in the oneline part of this commit.
The two can either ignore each other, move in the same
way, or move in the opposite way.  Fichtl cases
move in the same way.  Adamscases do correlation
or anticorrelation.  I have taken out flGaussdiffrand,
replacing that variable and its effects with chxsvartype.

I have also loaded the correct values for the Fichtl cases.