Density-based: KL_point correction
I had set xss to 0 when not using that cross section.
For lognormal this was not correct after taking the exponential.
Now I set to a negative number of large magnitude.

This commit also contains some of the next, performing or not
performing operations based on whether a cross section is used.