Numerical Cov: 'discrete' eigfunc evaluator
I would like to later add the options of linearinterp
and Nystrominterp, but if the grid is fine enough, I expect
this to have a negligable effect on the solution.  And these
interpolators are slower than the discrete version, so they
will probably not be the most used options for real problems,
leaving this eigenfunction evaluator to do the heavy lifting.

linearinterp in a test problem in python was about 2.5 times
longer in runtime than discrete.  Nystrom interp was about 75
times longer.

This appears to be working, producing the expected behavior
in Adams cases, Fichtl cases, and Olson cases.  The difference
from fitlamc to numeric was greatest when variances are large,
and though the differences are noticable, they are not extreme.