Simplify: Clean up directory
This commit addresses Issue #11.

It deletes some files that are no longer used.

It removes some files from being tracked by Git,
such as what have previously been pesky .mod binaries.
'git rm --cached file.mod' is a gem--the file is removed
from tracking but remains in the workspace.

It improves the .gitignore file significantly so that it is
organized, clean and effective.  Specifically, recursive and
glob referencing is used, and file types to be ignored are
compressed to an observable number of commands organized by
type with brief commenting on the type.