Numerical Cov: interpolation options added
I added the 'linear' and 'Nystrom' interpolation options.

The 'linear' option appears to be correct against the
'discrete' implementation.
The 'Nystrom' option seems off--there is likely a bug.
The python implementation does not square the average terms in
'Gausscov', or what here is 'GaussLNcov'.  I'm pretty sure they
should be squared.
Need to resolve both these issues, recorded in github as
Issues #12 and #13.

Along implementation, modified calls to Eigfunc, for backwards
compatibility, should be aware of this, i.e. Issue #8.