Numerical Cov: added input test
Added test to ensure that the Nystrom discretization chosen was
greater than or equal to the number of eigenvalues chosen for
each cross section being used.  If this is not true the python
script which solves the Fredholm equation using the Nystrom method
fails, causing the program to fail.  Now such a run is intercepted
in the input testing with a message about why it failed.