Simplify: consolidate 's','GB'->'slen', issue #16
I had a 'Markov-based' slab thickness 's' and a 'Gaussian-based'
slab thickness 'GBs' in the code.  There was no need to have
both, and in order to reduce the likelihood of accidentally
using the wrong slab thickness, I combined the two.  I chose
a new name which was not either of the pre-existing names
because 's' is hard to search for, it is frustratingly not
a unique letter combination, and 'GBs' specifically says that
it is the 'Gaussian-based' slab thickness whereas my new variable
name needs to work for both.  I thus chose 'slen', short for
'slab length', but still long enough to mostly be a unique,
searchable letter combination.