Numerical Cov: Analytic cov bug fix, issue #17
What I noticed was that the problem would not run with 4
processors with an analytic covariance.  I did three things
and now produce the expected results:

1) I tested 'lamctypeXX' for 'analytic', but that is not
an option, 'Glamc' or 'fitlamc' are, so I fixed that.
Without this change eigenvalues were never solved for.

2) Sometimes the 'zero' root of the Fredholm transcendental
equation was being found as the first root, when in fact
it is not the first valid root.  I added a hard value
of slightly higher than 0 for the starting point
in my searching algorithm.

3) I added a test of whether very small roots have been
found.  If they are the code stops and provides a comment.
This could always be modified if a true root was tripping
the assertion.