LN Proc: SC, old 'xi' debug
This variable is legacy from chord length sample generated
binary media.  I think I will use it again, however, so
it perseveres.  This one line fix allows MPI runs with a
different number of KL terms kept on different processes
to not throw an error because of the length of this
locally ubiquitous array.