LN Proc: SC, assymetric KL orders enabled, Issue #2 pt2, Issue #4
Several things were accomplished in this commit, with the
goal achievement being enabling stochastic collocation for
multiple correlations.

1) debug - I have set unused 'numEig' values to zero early
on.  This saved at least two issues with variables not
being used here, but around for other capabilities.

2) Allocate 'Qs' for multiple correaltions on slave processors
in the same was as on the master.

3) Allocate 'nodes', which reads nodes from the python cubature
generator according to 'Qs'.

4) Properly load these 'nodes' into 'xi' values for each process.
This was the end goal and the others needed to happen to get here.
This was the SC completion of Issue #2.

I did a few other small clean-up things as well.

I have checked that a Monte Carlo UQ approach and stochastic collocation
UQ approach converge to the same answer for several problems using the
new capabilities.  This gives me some confidence that the new
implemenation is functional.  This was Issue #4.