PCE: Add 'PCEcells', Isue #18
Added 'PCEcells', allocatable array which holds the flux
cells for which the PCE is applied.  Also created 'numPCEcells',
which is the size of this array.  Valid input include -1, 0, 1, 2, ...
Zero means no cells are chosen and negative one means the array
is to be made the same size as the number of flux cells.
If '-1' is chosen, the value of numPCEcells is changed to the
number of flux, and now PCE cells.

Several tests were added to make sure flux plotting is enabled,
and that the number of cells and cells chosen are valid.