PCE: Add PCEcoefs vars, utils, Issues #29,#20
I added 'numUQdims'.  This variable was able to replace 'size(Qs)' in
several places, and a whole do loop to recalculate the size Qs should be
for slave processors.

I added 'numPCEcoefs', which is the number of coefficients to be solved
for at a location in the slab.  To solve this value, I wrote a 'nCr'
function in the 'utilities.F90'.  I also added a 'factorial' function
while I was at it.  I solved simple inputs by hand to verify each
implementation, and used the two to verify each other for larger
numbers.

I added 'PCEcoefsrefl', 'PCEcoefstran', and 'PCEcoefscells', arrays
which are to hold the coefficients solved for at each chosen location
in the slab.

I added 'chUQtype=='PCE'' several places to logic which before only
occurred for 'SC', but needs to also for 'PCE'.