PCE: solve_PCEcoefs pt1, start, Issue #21
The first in a series of commits to implement solving of the
PCE coefficients from the problem solutions.  I decided to
create a subroutine which cycles through the PCE coefs to solve,
'solve_allPCEcoefs', which calls 'solve_PCEcoefs'.  The latter
accepts 'solns' from the first and passes 'coefs'.  These are
passed by reference for easy application of the subroutine
to PCE coefficients at different places in the slab.

The following commits are to be creating functions which this
calls, then finishing this subroutine.