PCE: solve_PCEcoefs pt2, UQnodes and cleanups
I added 'UQnodes', which will be useful in at least 'solve_PCEcoefs'.

In at least one place I added 'chUQtype=='PCE'' to the logic flow alongside
'chUQtype=='SC''.

I initialized and passed only 3 characters for 'chUQtype' instead of the
previous 5.

I added a deallocation of 'PCEpt' in the new 'solve_PCEcoefs' where it will
be needed.

I broadcast 'numPCEcoefs'.  It was solved before, but not broadcast.

I fixed slave logic initialization of 'Qs' to ensure that only slave logic
initializes to zero during the slave allocation phase.

I added the flag or equivalent integer conditions for 'PCEceofsXXXX' to
be allocated and initialized in slave logic.