PCE: solve_PCEcoefs, finish, Issue #25
I put the finishing touches on 'solve_PCEcoefs' for now.
This commit:

1) actually calls 'solve_allcoefs'

2) makes the dimension in the 'increment_PCEpt' function
an optional input so that the user doesn't have to pass
a '1'

3) removes one variable in the 'HermiteProbpoly' call,
as the third was not needed and thus incorrect

4) adds an if statement so that 'increment_PCEPt' is only
called after the first time through, so that the 0th moment
is solved for

5) when solving internal flux PCE coefs, passing based on
the PCE cell coef number, not the cell value.

These help to solidify this implementation.  This commit
includes some testing scaffolding as well.  The PCE coefs
in general seem somewhat reasonable.  The non-0th order
moments, however, do not seem to drop off as fast as I
would expect.  This should be investigated more when there
is a good way to benchmark such as seeing the final solution
that the coefficients produce and comparing to uncollided
flux benchmarks values in the python implementation.