BSD: use MT in two mods
I was having an issue which gave the error:
'
call init( mtx, int(rngappnum*rngstride_j,8))

Error: There is no specific subroutine for the generic 'init' at (1)
'
I needed to use a '4' instead of an '8' to meet the type
requirements needed for the init interface.

Here I replace the mcnp rng code with MT in 'genRealz.f90'
completely and initialize in 'Loadcase.f90'.